requirements.txt
asgiref==3.5.0
Django==4.0.4
django-cors-headers==3.12.0
django-restframework==0.0.1
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.1.0